requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
yfinance==0.2.28
alpha-vantage==2.3.1
openai==1.3.5
langchain==0.0.329
langchain-openai==0.0.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
plotly==5.17.0
streamlit-plotly-events==0.0.6
python-multipart==0.0.6
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1
